chuc nang dang ky api

+++ QLKhoHang/src/context/AuthContext.js
@@ -6,12 +6,13 @@ import { BASE_URL } from "../config";
 
 export const AuthContext = createContext();
 
-export const AuthProvider = ({ children}) => {
+export const AuthProvider = ({ children }) => {
   // const navigation = useNavigation();
 
   const [userInfo, setUserInfo] = useState({});
   const [isLoading, setIsLoading] = useState(false);
   const [splashLoading, setSplashLoading] = useState(false);
+  const [checkValue, setCheckValue] = useState(false);
 
   const signUP = (
     username,
@@ -21,28 +22,48 @@ export const AuthProvider = ({ children}) => {
     phone,
     email
   ) => {
+    // console.log(">>>>> check",username,password,confirmPassword,address,phone,email);
     setIsLoading(true);
-
+    // let acc = {
+    //   username,
+    //   password,
+    //   confirmPassword,
+    //   address,
+    //   email,
+    //   phone,
+    // };
+    console.log( username,
+      password,
+      confirmPassword,
+      address,
+      phone,
+      email);
     axios
       .post(`${BASE_URL}/register?status=2`, {
-        username,
-        password,
-        confirmPassword,
-        address,
-        phone,
-        email
+        username: username,
+        password: password,
+        confirmPassword: confirmPassword,
+        address: address,
+        email: email,
+        phone: phone,
       })
       .then((res) => {
+        console.log(res);
         let userInfo = res.data;
+        console.log(userInfo);
         AsyncStorage.setItem("userInfo", JSON.stringify(userInfo));
-        console.log(userInfo);
         setIsLoading(false);
-        console.log(userInfo);
       })
       .catch((e) => {
-        console.log(`register error ${e.response.data}`);
+        console.log(`register error ${e.res}`);
         setIsLoading(false);
       });
+      // console.log(`data `, username,
+      // password,
+      // confirmPassword,
+      // address,
+      // phone,
+      // email);
   };
 
   const login = (username, password) => {
@@ -57,11 +78,6 @@ export const AuthProvider = ({ children}) => {
         let userInfo = res.data;
         setUserInfo(userInfo);
         AsyncStorage.setItem("userInfo", JSON.stringify(userInfo));
-        // AsyncStorage.setItem("accessToken",userInfo.accessToken, JSON.stringify(userInfo.accessToken));
-        // AsyncStorage.setItem("message",userInfo.message, JSON.stringify(userInfo.message));
-        // AsyncStorage.setItem("others",userInfo.others, JSON.stringify(userInfo.others));
-        // setUserInfo(userInfo.accessToken);
-        // navigation.navigate('Home');
         setIsLoading(false);
       })
       .catch((e) => {
@@ -70,28 +86,28 @@ export const AuthProvider = ({ children}) => {
       });
   };
 
-  const logout = () => {
-    setIsLoading(true);
+  // const logout = () => {
+  //   setIsLoading(true);
 
-    axios
-      .post(
-        `${BASE_URL}/logout`,
-        {},
-        {
-          headers: { Authorization: `Bearer ${userInfo.access_token}` },
-        }
-      )
-      .then((res) => {
-        console.log(res.data);
-        AsyncStorage.removeItem("userInfo");
-        setUserInfo({});
-        setIsLoading(false);
-      })
-      .catch((e) => {
-        console.log(`logout error ${e}`);
-        setIsLoading(false);
-      });
-  };
+  //   axios
+  //     .post(
+  //       `${BASE_URL}/logout`,
+  //       {},
+  //       {
+  //         headers: { accessToken: `Bearer ${userInfo.accessToken}` },
+  //       }
+  //     )
+  //     .then((res) => {
+  //       console.log(res.data);
+  //       AsyncStorage.removeItem("userInfo");
+  //       setUserInfo({});
+  //       setIsLoading(false);
+  //     })
+  //     .catch((e) => {
+  //       console.log(`logout error ${e}`);
+  //       setIsLoading(false);
+  //     });
+  // };
 
   const isLoggedIn = async () => {
     try {
@@ -119,11 +135,11 @@ export const AuthProvider = ({ children}) => {
     <AuthContext.Provider
       value={{
         isLoading,
-        userInfo, 
+        userInfo,
         splashLoading,
         signUP,
         login,
-        logout,
+        // logout,
       }}
     >
       {children}
